refactor(service-datasource): merge federation + runtime admin into one package
Consolidate `@objectstack/service-external-datasource` (federation) and
`@objectstack/service-datasource-admin` (runtime UI lifecycle) into a single
`@objectstack/service-datasource`. The two were always one ADR-0015 capability;
they were only split because the runtime admin was once private. Now that both
are open-source that reason is gone, so they become one package.

The two plugins still mount independently (serve wires each in its own block) —
merging the package did not couple them. Plugin runtime IDs
(`com.objectstack.service-external-datasource` /
`com.objectstack.service-datasource-admin`) are unchanged, so dedup matching and
any persisted references keep working.

- git mv preserves history for both halves.
- Unified index re-exports both surfaces; single canonical `Logger`.
- package.json: driver packages as optional peers, exports `.` + `./contracts`,
  local tsup (index + contracts entries).
- Consumers updated: cli dep + serve.ts dynamic imports + .changeset/config.json.
- Now is the cheapest time to do this: only the cli depended on either package.

Verified: typecheck clean; 56 tests pass (federation + admin); package build
(incl. dts) and cli build green.
