+++ packages/cli/package.json
@@ -70,8 +70,7 @@
     "@objectstack/service-analytics": "workspace:*",
     "@objectstack/service-automation": "workspace:*",
     "@objectstack/service-cache": "workspace:*",
-    "@objectstack/service-datasource-admin": "workspace:*",
-    "@objectstack/service-external-datasource": "workspace:*",
+    "@objectstack/service-datasource": "workspace:*",
     "@objectstack/service-feed": "workspace:*",
     "@objectstack/service-job": "workspace:*",
     "@objectstack/service-messaging": "workspace:*",
