--- packages/services/service-external-datasource/package.json
@@ -1,47 +1,0 @@
-{
-  "name": "@objectstack/service-external-datasource",
-  "version": "7.5.0",
-  "license": "Apache-2.0",
-  "description": "External Datasource Federation service for ObjectStack — implements IExternalDatasourceService (introspect, draft, validate) per ADR-0015",
-  "type": "module",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "require": "./dist/index.cjs"
-    }
-  },
-  "scripts": {
-    "build": "tsup --config ../../../tsup.config.ts",
-    "test": "vitest run"
-  },
-  "dependencies": {
-    "@objectstack/core": "workspace:*",
-    "@objectstack/spec": "workspace:*"
-  },
-  "devDependencies": {
-    "@types/node": "^25.9.1",
-    "typescript": "^6.0.3",
-    "vitest": "^4.1.8"
-  },
-  "keywords": [
-    "objectstack",
-    "service",
-    "external-datasource",
-    "federation",
-    "introspection"
-  ],
-  "author": "ObjectStack",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/objectstack-ai/framework.git",
-    "directory": "packages/services/service-external-datasource"
-  },
-  "homepage": "https://objectstack.ai/docs",
-  "bugs": "https://github.com/objectstack-ai/framework/issues",
-  "publishConfig": {
-    "access": "public"
-  }
-}
